Be robust to dispatcher dropping in source removal
Dropping the dispatchers can run arbitrary `Drop` code written by the
user. This code can itself trigger the removal of other sources.

We thus delay the actual dropping of the dispatcher to *after* the
borrow on the SourceList is released, otherwise the abose case would
cause panics on double-borrows of the RefCell.