Add socket event source
This adds the standard Unix domain socket as a Calloop socket source.

While implementing this should be fairly easy for most calloop
consumers, it's also a pretty standard thing to need especially in the
context of event loops. So this patch should provide a helpful
abstraction saving some time downstream.