Use `io-lifetimes` types in API, instead of `RawFd`
Operating on `RawFd`s is unsafe, so this should be changed. Though some
uses will be cleaner once `nix` is updated for IO safety (which seems
to be ongoing, but complicated).

The `Rc<RefCell<IoDispatcher>>` in `Async` is somewhat awkward here. For
now `Dispatcher` still contains a `RawFd`. The only alternative I can
think of with how the API works is to use `Rc<F>` and have `into_inner`
call `Rc::try_unwrap(self.fd).unwrap()`.