break: Remove nix from the public API
This commit removes nix from the public API. nix has frequent breaking
changes due to the large API surface it covers, which makes it a semver
hazard to keep it in the public API.

Thus far only two items from nix are exposed: Signal and siginfo. Signal
is an enum consisting of most real-life Unix signals. I have replaced it
with a hand rolled enum primarily taken from the one in async-signal.
For the other case, the Event structure exposes most of the relevant
fields. So I've turned these fields into accessors. We can PR more
later.

This is a breaking change, but it's in service of avoiding breaking
changes in the future.

Closes #179

Signed-off-by: John Nunley <dev@notgull.net>