docs: Elaborate on the MSRV policy
Previously bumps in the MSRV were treated as breaking changes. However this is
unsustainable for a stable release, as some of our dependencies take bumps in
the MSRV as minor version bumps. This adds a section to the MSRV policy describing
what kind of version bumps we expect to take with regards to MSRV.

The policy is mostly copied from smol.

Signed-off-by: John Nunley <dev@notgull.net>