source/futures: Fix handling of `notified` atomic
Previously, `notified` was cleared before the
`self.source.process_events` call. This meant that it was cleared before
reading from the ping source eventfd/pipe.

This means that a call to `send()` form a waker in a different thread
could call `ping()` before this has been read, and prevent future calls
to `ping()`. But `process_events()` won't be woken again, freezing the
executor.

This seems to fix an issue I've seen in `cosmic-workspaces` when trying
to use the `calloop` executor.