Add `StreamSource` to adapt an async `Stream` to an `EventSource`
Async `Stream`s are logically quite similar an `EventSource`, though not
tied to fd polling. So it's quite natural to insert a stream as a
calloop source, and then have a callback handle any value yielded by the
stream. This behaves similarly to the `Channel` source.

Previously, this would require an executor, either the calloop one or an
executor on a different thread, running a future that iterates over the
stream and send return values on a calloop channel. This adapter event
source is simpler to use and more efficient, for cases where one or a
few streams need to be handled, but there's not a need to schedule very
large numbers of them, or run on multiple threads.